+++ src/exported.ts
@@ -108,7 +108,12 @@ export { PermissionSetAssignment, PermissionSetAssignmentFields } from './org/pe
 
 export { ScratchOrgCreateOptions, ScratchOrgCreateResult, scratchOrgCreate } from './org/scratchOrgCreate';
 
-export { ScratchOrgInfo } from './org/scratchOrgInfoApi';
+export { ScratchOrgInfo } from './org/scratchOrgTypes';
+export {
+  ScratchOrgLifecycleEvent,
+  scratchOrgLifecycleEventName,
+  scratchOrgLifecycleStages,
+} from './org/scratchOrgLifecycleEvents';
 
 // Utility sub-modules
 export * from './util/sfdc';